+++ src/propertyFields/dropdownWithCallout/PropertyFieldDropdownWithCalloutHost.tsx
@@ -1,6 +1,4 @@
 import * as React from 'react';
-import * as _ from 'lodash';
-
 import PropertyFieldHeader from '../../common/propertyFieldHeader/PropertyFieldHeader';
 
 import { IPropertyFieldDropdownWithCalloutHostProps } from './IPropertyFieldDropdownWithCalloutHost';
@@ -8,6 +6,8 @@ import * as telemetry from '../../common/telemetry';
 import { Dropdown, IDropdownProps, IDropdownOption } from 'office-ui-fabric-react/lib/components/Dropdown';
 import { IPropertyPaneDropdownOption } from '@microsoft/sp-webpart-base';
 import { SelectableOptionMenuItemType } from 'office-ui-fabric-react/lib/utilities/selectableOption/SelectableOption.types';
+
+const omit: any = require('lodash.omit');
 
 export default class PropertyFieldDropdownHost extends React.Component<IPropertyFieldDropdownWithCalloutHostProps, null> {
     constructor(props: IPropertyFieldDropdownWithCalloutHostProps) {
@@ -19,7 +19,7 @@ export default class PropertyFieldDropdownHost extends React.Component<IProperty
     }
 
     public render(): JSX.Element {
-      const dropdownProps: IDropdownProps = _.omit(this.props, ['label']);
+      const dropdownProps: IDropdownProps = omit(this.props, ['label']);
       dropdownProps.options = this._convertPropPaneOptionsToDropdownOptions(dropdownProps.options);
         return (
             <div>