+++ src/webparts/propertyControlsTest/components/PropertyControlsTest.tsx
@@ -7,8 +7,8 @@ import { escape } from '@microsoft/sp-lodash-subset';
  * Property control test component that renders the property control values
  */
 export default class PropertyControlsTest extends React.Component<IPropertyControlsTestProps, {}> {
-  public setHtml(){
-    return { __html:this.props.htmlCode };
+  public setHtml() {
+    return { __html: this.props.htmlCode };
   }
   public render(): React.ReactElement<IPropertyControlsTestProps> {
     return (
@@ -31,8 +31,8 @@ export default class PropertyControlsTest extends React.Component<IPropertyContr
               <p className="ms-font-m ms-fontColor-neutralDark">Multi Filtered: {this.props.multiListFiltered.join(', ')}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Term(s): {this.props.terms.map(t => t.name).join(', ')}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Date: {this.props.datetime.displayValue}</p>
-              <p className="ms-font-m ms-fontColor-neutralDark">Color: <span className={styles.colorBox} style={{backgroundColor:this.props.color}}>&nbsp;</span>{this.props.color}</p>
-              <p className="ms-font-m ms-fontColor-neutralDark">Color Object: <span className={styles.colorBox} style={{backgroundColor:this.props.colorObj ? this.props.colorObj.str : ''}}>&nbsp;</span>{this.props.colorObj ? `${this.props.colorObj.str}= R:${this.props.colorObj.r},G:${this.props.colorObj.g},B:${this.props.colorObj.b},A:${this.props.colorObj.a} H:${this.props.colorObj.h},S:${this.props.colorObj.s},V:${this.props.colorObj.v} HEX:${this.props.colorObj.hex}` : ''}</p>
+              <p className="ms-font-m ms-fontColor-neutralDark">Color: <span className={styles.colorBox} style={{ backgroundColor: this.props.color }}>&nbsp;</span>{this.props.color}</p>
+              <p className="ms-font-m ms-fontColor-neutralDark">Color Object: <span className={styles.colorBox} style={{ backgroundColor: this.props.colorObj ? this.props.colorObj.str : '' }}>&nbsp;</span>{this.props.colorObj ? `${this.props.colorObj.str}= R:${this.props.colorObj.r},G:${this.props.colorObj.g},B:${this.props.colorObj.b},A:${this.props.colorObj.a} H:${this.props.colorObj.h},S:${this.props.colorObj.s},V:${this.props.colorObj.v} HEX:${this.props.colorObj.hex}` : ''}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Spin Value: {this.props.spinValue}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Dropdown With Callout Key: {this.props.dropdownWithCalloutKey}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Slider With Callout Value: {this.props.sliderWithCalloutValue}</p>
@@ -41,7 +41,7 @@ export default class PropertyControlsTest extends React.Component<IPropertyContr
               <p className="ms-font-m ms-fontColor-neutralDark">Text Info Header Value: {this.props.textInfoHeaderValue}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Toggle Info Header Value: {this.props.toggleInfoHeaderValue ? 'Marvel' : 'DC Comics'}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Checkbox with Callout Value: {(this.props.checkboxWithCalloutValue || '').toString()}</p>
-              <p className="ms-font-m ms-fontColor-neutralDark" style={{wordBreak:"break-all"}}>Collection data: {JSON.stringify(this.props.collectionData)}</p>
+              <p className="ms-font-m ms-fontColor-neutralDark" style={{ wordBreak: "break-all" }}>Collection data: {JSON.stringify(this.props.collectionData)}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Ordered Items: {this.props.orderedItems.map((value: any) => {
                 return (
                   <i
@@ -51,8 +51,9 @@ export default class PropertyControlsTest extends React.Component<IPropertyContr
                   />
                 );
               })}</p>
-              <p className="ms-font-m ms-fontColor-neutralDark">Swatch Color: <span className={styles.colorBox} style={{backgroundColor:this.props.swatchColor}}>&nbsp;</span>{this.props.swatchColor}</p>
+              <p className="ms-font-m ms-fontColor-neutralDark">Swatch Color: <span className={styles.colorBox} style={{ backgroundColor: this.props.swatchColor }}>&nbsp;</span>{this.props.swatchColor}</p>
               <p className="ms-font-m ms-fontColor-neutralDark">Enterprise Term(s): {this.props.enterpriseTerms.map(t => t.name).join(', ')}</p>
+              <p className="ms-font-m ms-fontColor-neutralDark">Site(s): {this.props.sites.map(t => t.title).join(', ')}</p>
             </div>
           </div>
         </div>